NC件白名单

+++ jy-basic-api/src/main/java/com/jdl/basic/api/dto/ncWhiteList/NCWhiteRuleDTO.java
@@ -19,9 +19,9 @@ public class NCWhiteRuleDTO {
 
     private Integer lte;
 
-    private Integer gtValue;
+    private Double gtValue;
 
-    private Integer ltValue;
+    private Double ltValue;
 
     private String unit;
